services/price-service/pyproject.toml
# Price Service for PETMOL
# FastAPI backend that aggregates pet product prices from multiple sources

[project]
name = "petmol-price-service"
version = "0.1.0"
description = "Price comparison API for pet products"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "cachetools>=5.3.0",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.1.0",
    "openai>=1.0.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "pywebpush>=1.14.0",
    "apscheduler>=3.10.4",
    "google-generativeai>=0.8.0",
    "pypdf>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
